refactor: split Gitlab classes in separate files
