+++ plugins/gitlab/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@cpn-console/gitlab-plugin",
   "type": "module",
-  "version": "3.3.1",
+  "version": "3.4.0",
   "private": false,
   "description": "",
   "main": "dist/index.js",
