+++ plugins/gitlab/src/repositories.ts
@@ -2,7 +2,7 @@ import type { Project, Repository } from '@cpn-console/hooks'
 import type { VaultProjectApi } from '@cpn-console/vault-plugin/types/vault-project-api.js'
 import type { CondensedProjectSchema, ProjectSchema } from '@gitbeaker/rest'
 import { shallowEqual } from '@cpn-console/shared'
-import { type GitlabProjectApi, pluginManagedTopic } from './class.js'
+import { type GitlabProjectApi, pluginManagedTopic } from './gitlab-project-api.js'
 import { provisionMirror } from './project.js'
 import { infraAppsRepoName, internalMirrorRepoName } from './utils.js'
